refactor: reorganize schema files and update imports for improved structure
