+++ package.json
@@ -128,7 +128,7 @@
     "react-dom": "~19.2.1",
     "react-refresh": "~0.18.0",
     "rimraf": "~6.1.0",
-    "sass": "~1.98.0",
+    "sass": "~1.99.0",
     "sass-loader": "~16.0.6",
     "style-loader": "~4.0.0",
     "terser-webpack-plugin": "~5.4.0",
@@ -137,7 +137,7 @@
     "ts-node": "~10.9.2",
     "typescript": "~5.9.3",
     "typescript-eslint": "~8.58.0",
-    "webpack": "~5.105.0",
+    "webpack": "~5.106.0",
     "webpack-cli": "~7.0.2",
     "webpack-dev-server": "~5.2.2"
   },
@@ -179,7 +179,10 @@
       "diff": "~8.0.3"
     },
     "serialize-javascript": "~7.0.4",
-    "test-exclude": "~7.0.1"
+    "test-exclude": "~7.0.1",
+    "@microsoft/1ds-core-js": "~4.3.11",
+    "@microsoft/1ds-post-js": "~4.3.11",
+    "@microsoft/applicationinsights-web-basic": "~3.3.11"
   },
   "extensionDependencies": [
     "ms-azuretools.vscode-azureresourcegroups"
