+++ src/panels/QueryEditorTab.ts
@@ -11,10 +11,15 @@ import { type QuerySession } from '../cosmosdb/session/QuerySession';
 import { type SerializedQueryResult } from '../cosmosdb/types/queryResult';
 import { getIsSurveyDisabledGlobally } from '../utils/survey';
 import { TypedEventSink } from '../utils/TypedEventSink';
-import { type QueryEditorRouterContext } from '../webviews/api/configuration/appRouter';
-import { type QueryEditorEvent } from '../webviews/api/configuration/routers/queryEditorEventsRouter';
-import { setupTrpc } from '../webviews/api/extension-server/setupTrpc';
 import { BaseTab } from './BaseTab';
+import {
+    queryEditorAppRouter,
+    queryEditorCallerFactory,
+    type QueryEditorMutableState,
+    type QueryEditorRouterContext,
+} from './trpc/appRouter';
+import { type QueryEditorEvent } from './trpc/routers/queryEditorEventsRouter';
+import { setupTrpc } from './trpc/setupTrpc';
 
 export class QueryEditorTab extends BaseTab {
     public static readonly title = 'Query Editor';
@@ -24,22 +29,22 @@ export class QueryEditorTab extends BaseTab {
     public readonly sessions = new Map<string, QuerySession>();
     public readonly eventSink: TypedEventSink<QueryEditorEvent>;
 
-    private connection: NoSqlQueryConnection | undefined;
-    private query: string | undefined;
-    private isLastQueryAIGenerated: boolean = false;
-    /** The raw AI-generated query text, used to detect if the user modified it before running */
-    private lastAIGeneratedQuery: string | undefined;
-    private lastGenerationFailed: boolean = false;
-    private generateQueryCancellation: vscode.CancellationTokenSource | undefined;
-    private pendingConfirmResolve: ((confirmed: boolean) => void) | undefined;
+    private readonly state: QueryEditorMutableState;
 
     protected constructor(panel: vscode.WebviewPanel, connection?: NoSqlQueryConnection, query?: string) {
         super(panel, QueryEditorTab.viewType, { hasConnection: connection ? 'true' : 'false' });
 
         QueryEditorTab.openTabs.add(this);
 
-        this.connection = connection;
-        this.query = query;
+        this.state = {
+            connection,
+            query,
+            isLastQueryAIGenerated: false,
+            lastAIGeneratedQuery: undefined,
+            lastGenerationFailed: false,
+            generateQueryCancellation: undefined,
+            pendingConfirmResolve: undefined,
+        };
 
         this.panel.iconPath = getThemedIconPath('editor.svg') as { light: vscode.Uri; dark: vscode.Uri };
 
@@ -58,7 +63,12 @@ export class QueryEditorTab extends BaseTab {
         // Create TypedEventSink for tRPC subscription
         this.eventSink = new TypedEventSink<QueryEditorEvent>();
 
-        const { disposable } = setupTrpc(this.panel, this.buildRouterContext());
+        const { disposable } = setupTrpc(
+            this.panel,
+            this.buildRouterContext(),
+            queryEditorAppRouter,
+            queryEditorCallerFactory,
+        );
         this.disposables.push(disposable);
     }
 
@@ -71,9 +81,9 @@ export class QueryEditorTab extends BaseTab {
         if (revealTabIfExist && connection) {
             const openTab = [...QueryEditorTab.openTabs].find(
                 (openTab) =>
-                    openTab.connection?.endpoint === connection.endpoint &&
-                    openTab.connection?.databaseId === connection.databaseId &&
-                    openTab.connection?.containerId === connection.containerId,
+                    openTab.state.connection?.endpoint === connection.endpoint &&
+                    openTab.state.connection?.databaseId === connection.databaseId &&
+                    openTab.state.connection?.containerId === connection.containerId,
             );
             if (openTab) {
                 openTab.panel.reveal(viewColumn);
@@ -101,58 +111,13 @@ export class QueryEditorTab extends BaseTab {
     }
 
     private buildRouterContext(): QueryEditorRouterContext {
-        // We need `self` because getter/setter `this` inside the object literal
-        // refers to the routerContext object, not the QueryEditorTab instance.
-        // eslint-disable-next-line @typescript-eslint/no-this-alias
-        const self = this;
         return {
             webviewName: QueryEditorTab.viewType,
-            get connection() {
-                return self.connection;
-            },
             sessions: this.sessions,
             telemetryContext: this.telemetryContext,
             panel: this.panel,
             eventSink: this.eventSink,
-            get query() {
-                return self.query;
-            },
-            set query(v) {
-                self.query = v;
-            },
-            get isLastQueryAIGenerated() {
-                return self.isLastQueryAIGenerated;
-            },
-            set isLastQueryAIGenerated(v) {
-                self.isLastQueryAIGenerated = v;
-            },
-            get lastAIGeneratedQuery() {
-                return self.lastAIGeneratedQuery;
-            },
-            set lastAIGeneratedQuery(v) {
-                self.lastAIGeneratedQuery = v;
-            },
-            get lastGenerationFailed() {
-                return self.lastGenerationFailed;
-            },
-            set lastGenerationFailed(v) {
-                self.lastGenerationFailed = v;
-            },
-            get generateQueryCancellation() {
-                return self.generateQueryCancellation;
-            },
-            set generateQueryCancellation(v) {
-                self.generateQueryCancellation = v;
-            },
-            get pendingConfirmResolve() {
-                return self.pendingConfirmResolve;
-            },
-            set pendingConfirmResolve(v) {
-                self.pendingConfirmResolve = v;
-            },
-            setConnection: (conn) => {
-                self.connection = conn;
-            },
+            state: this.state,
         };
     }
 
@@ -163,11 +128,11 @@ export class QueryEditorTab extends BaseTab {
     };
 
     public getConnection = (): NoSqlQueryConnection | undefined => {
-        return this.connection;
+        return this.state.connection;
     };
 
     public getCurrentQuery = (): string | undefined => {
-        return this.query;
+        return this.state.query;
     };
 
     public isActive(): boolean {
@@ -188,9 +153,9 @@ export class QueryEditorTab extends BaseTab {
     }
 
     public async updateQuery(query: string): Promise<void> {
-        this.query = query;
-        this.isLastQueryAIGenerated = true;
-        this.lastAIGeneratedQuery = query;
+        this.state.query = query;
+        this.state.isLastQueryAIGenerated = true;
+        this.state.lastAIGeneratedQuery = query;
         this.eventSink.emit({ type: 'queryTextPushed', query });
     }
