+++ src/panels/trpc/schemas/aiSchemas.ts
