+++ src/webviews/api/trpc/vscodeLink.ts
@@ -4,44 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { TRPCClientError, type Operation, type TRPCLink } from '@trpc/client';
-// eslint-disable-next-line import/no-internal-modules
+import { type AnyRouter } from '@trpc/server';
 import { observable } from '@trpc/server/observable'; // Their example uses a reference from /server/ and so do we: https://trpc.io/docs/client/links#example
-import { type AppRouter } from '../configuration/appRouter';
+import { type VsCodeLinkRequestMessage, type VsCodeLinkResponseMessage } from '../../../panels/trpc/vscodeProtocol';
 
-type StopOperation<TInput = unknown> = Omit<Operation<TInput>, 'type'> & {
-    type: 'subscription.stop';
-};
-
-/**
- * Messages sent from the webview/client to the extension/server.
- * @id - A unique identifier for the message/
- */
-export interface VsCodeLinkRequestMessage {
-    id: string;
-    // TODO, when tRPC v12 is released, 'subscription.stop' should be supported natively, until then, we're adding it manually.
-    op: Operation<unknown> | StopOperation<unknown>;
-}
-
-/**
- * Messages sent back from the extension/server to the webview/client.
- * Each message sent back is a **response** to a previous message VsCodeLinkRequestMessage
- *
- * @id - The unique identifier of the message from the original request
- */
-export interface VsCodeLinkResponseMessage {
-    id: string;
-    result?: unknown;
-    error?: {
-        name: string;
-        message: string;
-
-        code?: number;
-        stack?: string;
-        cause?: unknown;
-        data?: unknown;
-    };
-    complete?: boolean;
-}
+export type { VsCodeLinkRequestMessage, VsCodeLinkResponseMessage } from '../../../panels/trpc/vscodeProtocol';
 
 interface VSCodeLinkOptions {
     //   Function to send a message to the server / extension
@@ -56,7 +23,7 @@ interface VSCodeLinkOptions {
  * @param options - The options for the link, including send and onReceive functions.
  * @returns A tRPC link for client-side usage.
  */
-function vscodeLink(options: VSCodeLinkOptions): TRPCLink<AppRouter> {
+function vscodeLink<TRouter extends AnyRouter>(options: VSCodeLinkOptions): TRPCLink<TRouter> {
     const { send, onReceive } = options;
 
     /**
